backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
gunicorn==23.0.0
sqlalchemy==2.0.40
python-dotenv==1.1.0
python-multipart==0.0.20
httpx==0.28.1
pandas==2.2.3
openpyxl==3.1.5
psycopg2-binary==2.9.10
